Convert the User/Me pages to use PageRoutedTabs. Add Back to Users tab.
